chore(web): move enums out of .svelte file
